requirements.txt
aiohttp==3.6.2
aiosqlite==0.15.0
async-timeout==3.0.1
attrs==20.1.0
chardet==3.0.4
discord.py==1.4.1
idna==2.10
multidict==4.7.6
typing-extensions==3.7.4.3
# -e git+https://github.com/jackodsteel/ValorantCompetitiveBot.git@642a9e25459ac5d4b21b1eaaee04c5485775fef3#egg=valorantcompetitivebot
# yarl==1.5.1
